requirements.txt
tensorflow-macos
tensorflow-metal
mlflow
fastapi
uvicorn
pillow
scikit-learn
python-multipart
